# Conflicts:
#	dynamokt/src/main/kotlin/com/steamstreet/dynamokt/DynamoKtSession.kt
#	dynamokt/src/main/kotlin/com/steamstreet/dynamokt/ItemContainer.kt
#	dynamokt/src/main/kotlin/com/steamstreet/dynamokt/ItemUpdater.kt
#	dynamokt/src/main/kotlin/com/steamstreet/dynamokt/MutableItem.kt
#	env/build.gradle.kts
#	standards/build.gradle.kts
#	standards/src/commonMain/kotlin/com/steamstreet/MutableLazy.kt
#	test/src/jvmMain/kotlin/com/steamstreet/aws/test/dynamo.kt

+++ test/src/jvmMain/kotlin/com/steamstreet/aws/test/dynamo.kt
@@ -281,7 +281,7 @@ class PipesConfiguration(
             record.awsRegion,
             record.dynamodb.let {
                 DynamoStreamEventDetail(
-                    it.approximateCreationDateTime.time,
+                    it.approximateCreationDateTime.time.toDouble(),
                     srcRecord.dynamodb!!.keys!!.toModelAttributeValue(),
                     srcRecord.dynamodb?.newImage?.toModelAttributeValue(),
                     srcRecord.dynamodb?.oldImage?.toModelAttributeValue()
